Rename 'bus' package to 'eventbus'
