+++ eventbus/event_bus.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bus
+package eventbus
 
 import asaskevichEventBus "github.com/asaskevich/EventBus"
